+++ blog/2024-09-14-expandable/2024-09-14-expandable.mdx
@@ -9,7 +9,7 @@ tags:
     - customComponents
     - react
 ---
-import Expandable from '@site/src/components/Expandable'
+import Expandable from '@site/src/components/Expandable/Expandable'
 
 I really like Docusaurus, but I also find there are some things it's missing.
 Happily, one of the _best_ things about it is that it is, at bottom, React,
