+++ docs/projects/projects.mdx
@@ -1,8 +1,6 @@
 ---
 title: Projects
 ---
-import Expandable from '@site/src/components/Expandable';
-
 ## My Projects
 
 Placeholder --- substantial content pending
