+++ src/components/Expandable/Expandable.stories.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import type { Meta, StoryObj } from '@storybook/react';
-import Expandable from '../../../components/Expandable';
-import type { ExpandablePropsType } from '../../../components/Expandable';
-import Lorem from '../../../../.storybook/assets/mdx/lorem_ipsum.mdx';
+import Expandable from './Expandable';
+import type { ExpandablePropsType } from './Expandable';
+import Lorem from '/.storybook/assets/mdx/lorem_ipsum.mdx';
 
 const render = (args: ExpandablePropsType) => (
     <>
